pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.32.1",
    "langgraph==0.3.5",
    "langchain-anthropic==0.3.3",
    "langchain-openai==0.3.0",
    "langchain-google-genai==2.0.7",
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    "asyncpg==0.30.0",
    "redis==5.2.1",
    "pinecone==5.4.2",
    "tavily-python==0.5.1",
    "playwright==1.49.1",
    "beautifulsoup4==4.12.3",
    "grpcio==1.68.1",
    "grpcio-tools==1.68.1",
    "structlog==24.4.0",
    "prometheus-client==0.21.1",
    "slowapi==0.1.9",
    "python-jose[cryptography]==3.3.0",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "ruff",
    "mypy",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
